agent-core/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
sqlalchemy==2.0.25
alembic==1.13.1
openai==1.12.0
anthropic==0.18.0
gitpython==3.1.41
pygithub==2.1.1
python-gitlab==4.4.0
python-jenkins==1.8.2
aiohttp==3.9.1
loguru==0.7.2
tenacity==8.2.3
jinja2==3.1.3

ollama==0.1.6
groq==0.4.1

google-generativeai==0.3.2
